+++ src/SteamScrapper.Domain/Services/Contracts/SubData.cs
@@ -1,4 +1,4 @@
 ﻿namespace SteamScrapper.Domain.Services.Contracts
 {
-    public record SubData(int SubId, string Title);
+    public record SubData(long SubId, string Title);
 }
